requirements.txt
flask
numpy
pandas
openai
faiss-cpu
Flask-Cors
matplotlib
plotly
scipy
scikit-learn
gunicorn